all fun has been finished

+++ TS70_DWCodeA/Emembed/Inc/modbus.h
@@ -13,12 +13,12 @@
 
 typedef struct 
 {
-    uint8_t  byte_info_H;                //DATA_H
-    uint8_t  byte_info_L;                //DATA_L
-    uint8_t  byte_cnt;                   //BYTE NUM*2
-    uint8_t  send_val_addr;            //DATA1 H 位置
-    uint8_t  rcv_value_addr;             //DATA1 H 位置
-    uint16_t start_addr;                 //查询起始位置
+    uint8_t  byte_info_H;               //DATA_H
+    uint8_t  byte_info_L;               //DATA_L
+    uint8_t  byte_cnt;                  //BYTE NUM*2
+    uint16_t  send_val_addr;            //DATA1 H 位置
+    uint16_t  rcv_value_addr;           //DATA1 H 位置
+    uint16_t start_addr;                //查询起始位置
 }MODBIS5_INFO;
 
 
@@ -31,7 +31,7 @@ void Modbus_Fun03( void );
 void Modbus_Fun06( void );
 void Modbus_Fun16( void );
 void slave_to_master(uint8_t code_num,uint8_t length);
-uint16_t MODBUS_CRC16(uint8_t *buf, uint8_t length);
+uint16_t MODBUS_CRC16(uint8_t *buf, uint16_t length);
 void write_slave_06( uint8_t uart_num, uint16_t reg_addr, uint8_t reg_val_H, uint8_t reg_val_L);
 
 #endif
